Renamed socket to GameServer
